+++ packages/hardhat/contracts/Contract.sol
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity =0.7.6;
+
+contract Contract {}
