+++ packages/hardhat/test/foundry-test/Contract.t.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity =0.7.6;
+
+import "forge-std/Test.sol";
+
+contract ContractTest is Test {
+    function setUp() public {}
+
+    function testExample() public {
+        assertTrue(true);
+    }
+}
