+++ packages/hardhat/test/hardhat-test/integration-tests/oracle-attack.ts
@@ -3,7 +3,7 @@ import BigNumberJs from 'bignumber.js'
 import { BigNumber } from "ethers";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
 import { expect } from "chai";
-import { Controller, INonfungiblePositionManager, MockErc20, ShortPowerPerp, WETH9, WPowerPerp, IUniswapV3Pool, ISwapRouter } from "../../typechain";
+import { Controller, INonfungiblePositionManager, MockErc20, ShortPowerPerp, WETH9, WPowerPerp, IUniswapV3Pool, ISwapRouter } from "../../../typechain";
 import { deployUniswapV3, deploySqueethCoreContracts, deployWETHAndDai, addSqueethLiquidity, addWethDaiLiquidity } from '../setup'
 import { isSimilar, getNow, one, oracleScaleFactor } from "../utils";
